res gated fix

+++ nlns/models/res_gated_gcn.py
@@ -190,13 +190,13 @@ class ResGatedGCN(pl.LightningModule):
         if self.compute_weights:
             np_y = data.y.cpu().numpy()
             cw = compute_class_weight("balanced", classes=np.unique(np_y), y=np_y)
-            weights = torch.tensor(cw).to(pred.device)
+            weights = torch.tensor(cw).to(pred.device).float()
 
         # turn pred into 2 category loss: edge in the final solution or not
         pred = torch.cat((pred, 1 - pred), axis=1)
         y = torch.cat((data.y.reshape(-1, 1), 1 - data.y.reshape(-1, 1)), axis=1)
 
-        loss = nn.functional.binary_cross_entropy(pred, y, weight=weights.float())
+        loss = nn.functional.binary_cross_entropy(pred, y, weight=weights)
         return pred_adj, loss
 
     def predict(self, data: Batch) -> Tuple[torch.Tensor, nn.BCELoss]:
